Changed return type of Type::bit_size (and overrides) to be size_t instead of unsigned short; now that array types exist, the bit_size could be much more than would fit into a 16-bit number; fixed typo in error message for array typechecking

+++ src/ast.cpp
@@ -27,7 +27,7 @@ void RangeType::print(std::ostream& output) const
     output << "Type: " << name << " Range: " << range;
 }
 
-unsigned short ArrayType::bit_size() const
+size_t ArrayType::bit_size() const
 {
     // TODO: account for padding of elements
     return index_type->range.size() * element_type->bit_size();
